+++ lib/src/screen.dart
@@ -1,7 +1,9 @@
 /// The default parameters of the screen of the aRdent device
 class GYWScreenParameters {
+  GYWScreenParameters._();
+
   /// Supported Width of a GYW aRdent device
-  static const int width = 800;
+  static const int width = 854;
 
   /// Supported Width of a GYW aRdent device
   static const int height = 480;